pyproject.toml
[tool.poetry]
name = "medical-report-analysis"
version = "0.1.0"
description = "A Streamlit app for analyzing medical reports and providing comprehensive health recommendations."
authors = ["Devi Charan <your-email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.10.0"
PyPDF2 = "^3.0.0"
python-dotenv = "^1.0.0"
crewai = "^0.2.0"
ollama = "^0.4.4"
langchain-community = "^0.3.0"

[tool.poetry.scripts]
medical-report-analysis = "app:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
